♻️[Refactor] 공통 레이아웃 컴포넌트 HeaderSection 버튼 prop으로 변경 #157
